requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
blinker==1.6.2
CacheControl==0.12.11
cachelib==0.10.2
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
DateTime==5.5
Deprecated==1.2.13
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.2.2
firebase==4.0.1
firebase-admin==6.1.0
Flask==2.3.2
Flask-Session==0.5.0
gcloud==0.18.3
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.168.0
google-auth==2.18.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
groq==0.23.1
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.2
msgpack==1.0.5
numpy==2.2.5
oauth2client==4.1.3
oauthlib==3.2.2
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pillow==11.2.1
proto-plus==1.22.2
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.7.0
pyparsing==3.0.9
PyPDF2==3.0.1
Pyrebase4==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jwt==4.0.0
pytz==2025.2
requests==2.30.0
requests-oauthlib==2.0.0
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.3.4
wrapt==1.15.0
zipp==3.15.0
zope.interface==7.2
